+++ Kikurage/App/Navigation/NavigationProtocol+Communication.swift
@@ -9,20 +9,10 @@
 import UIKit
 
 protocol CommunicationAccessable: PushNavigationProtocol, SafariViewNavigationProtocol {
-    func pushToCommunication()
     func presentToSafariView(from vc: UIViewController, urlString: String?, onError: (() -> Void)?)
 }
 
 extension CommunicationAccessable {
-    // MARK: - Push
-
-    func pushToCommunication() {
-        guard let vc = R.storyboard.communicationViewController.instantiateInitialViewController() else {
-            return
-        }
-        push(to: vc)
-    }
-
     // MARK: - SafariView
 
     func presentToSafariView(from vc: UIViewController, urlString: String?, onError: (() -> Void)?) {
