+++ Kikurage/App/Navigation/NavigationProtocol+Cultivation.swift
@@ -9,20 +9,12 @@
 import UIKit
 
 protocol CultivationAccessable: PushNavigationProtocol, ModalNavigationProtocol {
-    func pushToCultivation()
     func pushToCultivationDetail(cultivation: KikurageCultivation)
     func modalToPostCultivation()
 }
 
 extension CultivationAccessable {
     // MARK: - Push
-
-    func pushToCultivation() {
-        guard let vc = R.storyboard.cultivationViewController.instantiateInitialViewController() else {
-            return
-        }
-        push(to: vc)
-    }
 
     func pushToCultivationDetail(cultivation: KikurageCultivation) {
         guard let vc = R.storyboard.cultivationDetailViewController.instantiateInitialViewController() else {
