+++ Kikurage/App/Navigation/NavigationProtocol+Recipe.swift
@@ -9,20 +9,10 @@
 import UIKit
 
 protocol RecipeAccessable: PushNavigationProtocol, ModalNavigationProtocol {
-    func pushToRecipe()
     func modalToPostRecipe()
 }
 
 extension RecipeAccessable {
-    // MARK: - Push
-
-    func pushToRecipe() {
-        guard let vc = R.storyboard.recipeViewController.instantiateInitialViewController() else {
-            return
-        }
-        push(to: vc)
-    }
-
     // MARK: - Modal
 
     func modalToPostRecipe() {
